feat(devtools): extract devtools as module
